+++ client/src/stores/RoleMixin.js
@@ -1,13 +1,13 @@
 // mixins/adminMixin.js
 import { computed } from 'vue';
 import { useTokenStore } from '@/stores/TokenStore';
-import { useAdminStore } from '@/stores/AdminStore';
+import { useReviewerStore } from '@/stores/ReviewerStore';
 import { storeToRefs } from 'pinia';
 
 export default {
   setup() {
     // Stores
-    const adminStore = useAdminStore();
+    const adminStore = useReviewerStore();
     const tokenStore = useTokenStore();
 
     // Store states
